+++ shuup/utils/fields.py
@@ -29,3 +29,20 @@ class RelaxedModelChoiceField(forms.ModelChoiceField):
                             if obj is None or isinstance(obj, self.queryset.model):
                                 return obj
             raise verr  # Just reraise the original exception then, but from here for clarity
+
+
+class TypedMultipleChoiceWithLimitField(forms.TypedMultipleChoiceField):
+    def __init__(self, min_limit=None, max_limit=None, **kwargs):
+        self.min_limit = min_limit
+        self.max_limit = max_limit
+        super().__init__(**kwargs)
+
+    def clean(self, value):
+        value = super().clean(value)
+        if self.min_limit is not None and len(value) < self.min_limit:
+            ending = "" if self.min_limit % 10 == 1 else "s"
+            raise forms.ValidationError(f"You can't select less than {self.min_limit} item{ending}.")
+        if self.max_limit is not None and len(value) > self.max_limit:
+            ending = "" if self.max_limit % 10 == 1 else "s"
+            raise forms.ValidationError(f"You can't select more than {self.max_limit} item{ending}.")
+        return value