+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+### Added
+
+- Admin: New type of attribute: CHOICES
+- Admin: Selection of multiple choices from attribute for product
+- Front: Product filter by attribute choices
+
 ### Changed
 
 - Pull strings from Transifex